+++ CHANGELOG.md
@@ -6,12 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Fixed
+ - Avoid `ValueError: wrapper has not been initialized` when using implicit referential lambda fixtures (e.g. `name = lambda_fixture()`) in combination with `py.test --doctest-modules`
 
 
 ## [1.2.4] — 2020-12-28
 ### Changed
-- Add support for pytest version 6.2
-- Relax pytest version pin to allow all versions under 7.x
+ - Add support for pytest version 6.2
+ - Relax pytest version pin to allow all versions under 7.x
 
 
 ## [1.2.3] — 2020-11-02